refactor: 抽取 FullAccessWarning 独立组件，减轻 MainChat 职责
- 全权限敏感工具黄字警告从 MainChat.vue 移至 components/tools/FullAccessWarning.vue
- 工具设置预加载、watch 逻辑与样式随组件内聚，MainChat 仅挂载
- scoped 样式隔离，Transition/aria 语义保持不变
